+++ agent/check_handler_internal_test.go
@@ -14,6 +14,37 @@ import (
 var binDir = filepath.Join("..", "bin")
 var toolsDir = filepath.Join(binDir, "tools")
 
+func TestHandleCheck(t *testing.T) {
+	assert := assert.New(t)
+
+	checkConfig := types.FixtureCheckConfig("check")
+	truePath := testutil.CommandPath(filepath.Join(toolsDir, "true"))
+	checkConfig.Command = truePath
+
+	request := &types.CheckRequest{Config: checkConfig}
+	payload, err := json.Marshal(request)
+	if err != nil {
+		assert.FailNow("error marshaling check request")
+	}
+
+	config := NewConfig()
+	agent := NewAgent(config)
+	ch := make(chan *transport.Message, 5)
+	agent.sendq = ch
+
+	// check is already in progress, it shouldn't execute
+	agent.inProgressMu.Lock()
+	agent.inProgress[request.Config.Name] = request.Config
+	agent.inProgressMu.Unlock()
+	assert.Error(agent.handleCheck(payload))
+
+	// check is not in progress, it should execute
+	agent.inProgressMu.Lock()
+	delete(agent.inProgress, request.Config.Name)
+	agent.inProgressMu.Unlock()
+	assert.NoError(agent.handleCheck(payload))
+}
+
 func TestExecuteCheck(t *testing.T) {
 	assert := assert.New(t)
 
@@ -28,6 +59,7 @@ func TestExecuteCheck(t *testing.T) {
 
 	truePath := testutil.CommandPath(filepath.Join(toolsDir, "true"))
 	checkConfig.Command = truePath
+	checkConfig.Timeout = 10
 
 	agent.executeCheck(request)
 
@@ -36,7 +68,7 @@ func TestExecuteCheck(t *testing.T) {
 	event := &types.Event{}
 	assert.NoError(json.Unmarshal(msg.Payload, event))
 	assert.NotZero(event.Timestamp)
-	assert.EqualValues(0, event.Check.Status)
+	assert.EqualValues(int32(0), event.Check.Status)
 
 	falsePath := testutil.CommandPath(filepath.Join(toolsDir, "false"))
 	checkConfig.Command = falsePath
@@ -48,7 +80,20 @@ func TestExecuteCheck(t *testing.T) {
 	event = &types.Event{}
 	assert.NoError(json.Unmarshal(msg.Payload, event))
 	assert.NotZero(event.Timestamp)
-	assert.EqualValues(1, event.Check.Status)
+	assert.EqualValues(int32(1), event.Check.Status)
+
+	sleepPath := testutil.CommandPath(filepath.Join(toolsDir, "sleep"), "5")
+	checkConfig.Command = sleepPath
+	checkConfig.Timeout = 1
+
+	agent.executeCheck(request)
+
+	msg = <-ch
+
+	event = &types.Event{}
+	assert.NoError(json.Unmarshal(msg.Payload, event))
+	assert.NotZero(event.Timestamp)
+	assert.EqualValues(int32(2), event.Check.Status)
 }
 
 func TestPrepareCheck(t *testing.T) {